requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
python-dotenv
sqlalchemy
bcrypt
PyJWT
openai
anthropic
google-genai
requests
pypdf
python-docx
textstat
nltk
rouge_score
scikit-learn
keybert
